chore: moved component out of card folder
