+++ mobile-app/components/common/restaurantCard/RestaurantCard.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Image, View } from "react-native";
 import { Card, Text } from "react-native-paper";
-import { icons } from "../../../../constants";
+import { icons } from "../../../constants";
 import styles from "./restaurantCard.style";
 
 const RestaurantCard = ({ data, handlePress }) => {
